Create constants, move interface to types.ts
